rename CognitiveError to ClearableText
